event_user_onboarding_service/requirements.txt
fastapi==0.110.0
uvicorn==0.17.6
SQLAlchemy==1.4.35
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pika==1.3.2